requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pillow
scipy